[Fix #157370148] Refactor frontend and improve app User Experience
1. Integrate font-awesome to handle submit processing feedback
2. Replace alerts with toastr
3. Discard component validate method and use service userValidator
4. Refactor table.js component implementation to use .map method, use random key and eliminate colNumber prop
5. Remove unnecessary content variables in component render methods
6. Validate profile data onSubmit
7. Add object-fit css in .profile-image class
[Finishes #157370148]

+++ server/controllers/facilityController.js
@@ -15,17 +15,20 @@ class Facilities {
       centerId,
     }));
     return models.Centers.findOne({ where: { userId: req.decoded.id, id: centerId } })
-      .then(() =>
-        model.destroy({ where: { centerId } }).then(() =>
-          model.bulkCreate(newFacilities, { validate: true }).then(() =>
-            model
-              .findAll()
-              .then(facilities => validator.response(res, 'success', 200, facilities))
-              .catch(error => validator.response(res, 'error', 500, error))
-          )
-        )
-      )
-      .catch(() => validator.response(res, 'error', 403, 'Unauthorized transaction'));
+      .then(found => {
+        if (found) {
+          return model.destroy({ where: { centerId } }).then(() =>
+            model.bulkCreate(newFacilities, { validate: true }).then(() =>
+              model
+                .findAll()
+                .then(facilities => validator.response(res, 'success', 200, facilities))
+                .catch(error => validator.response(res, 'error', 500, error))
+            )
+          );
+        }
+        return validator.response(res, 'error', 403, 'Unauthorized transaction');
+      })
+      .catch(error => validator.response(res, 'error', 500, error));
   }
 }
 const facilityController = new Facilities();
